refactor(matching): address CodeRabbit review for PR #735
- Extract MatchingVoteContent to widgets/matching_vote_content.dart
  for better file separation and reusability
- Add explicit error state handling for vote meta providers
  (voteMetaError branch shows error message + disables buttons)
- Add test for vote meta error state
- Re-export MatchingVoteContent from matching_vote_screen.dart
  for backward compatibility

Closes #690
